Move waitForDomElement into page API and guard it against page reloads.
When a page gets reloaded, or a link gets triggered, then the
JavaScript context of the injectJS call gets invalidated. This
then stops the Promise of waitForDomElement, never letting it
finish.

Now, the code is evaluated in the PhantomJS context. Additionally,
some of the code is made more resilient against returning DOM
elements from cross-site iframes.