Reliably handle OnLoadEnd for non-HTTP requests.
Those pages have a zero http status code, just like the initial
about:blank load. We now add a boolean to the BrowserInfo to keep
track of the initial load and only propagate the load end for
subsequent requests.

This is exemplified by an onReourceError example, which shows that
we don't get any events when images or scripts get loaded but
then fail.