Add support for sending mouse move events.
This is a first step towards dragevent support, but that is
not yet fully functional. The drag event is fired, but no drop or
or dragover events are triggered yet.

Additionally this patch fixes an issue in the serialization of
arrays containing unserializable data. We now manually loop over
arrays and build a proper copy to remidy that.