Add page.onDownloadUpdated and page.onBeforeDownload signals.
The former can be used to query the status for an ongoing download
or cancel the download. The latter can be used to set the target
location of the downloaded file.

Triggering a download via page.download now makes the final data
accessible in the continuation callback. For other downloads triggered
by manual user interaction we now also get a download signal.